Prefix to postfix
The infixApp uses a stack to experiment a prefix expression and converts to a postfix. In addition, the program also asks the user to insert values for the variables in the Postfix, and evaluates the expression.
THE ALGORITHM USED IS AS FOLLOWS:
Operand: Write it to output (postfix)
Open parenthesis: ( Push it on stack
Close parenthesis ): While stack not empty, repeat the following:
Pop an item,
If item is not (, write it to output
Quit loop if item is (
Operator (opThis) If stack empty,
Push opThis
Otherwise,
While stack not empty, repeat:
Pop an item,
If item is (, push it, or
If item is an operator (opTop), and
If opTop < opThis, push opTop, or
If opTop >= opThis, output opTop
Quit loop if opTop < opThis or item is (
Push opThis
No more items While stack not empty,
Pop item, output it.